feature: Alternative modification to Rotation() instead of Right Handed and Left Handed interconversions
It seems that getting a generalized conversion model from left handed coordinate system to right handed and vice versa won't work since the order in which euler angles are visualized might matter in this case. Therefore, until further confirmations are made regarding carla's geometry handling, as an alternative the coordinate system flag has been introduced into the Rotation object; because the only parameter being changed currently is the right direction vector due the y-axis being flipped for the left handed system, related attributes can make use of the flag for distinguishing between a right handed and left handed system. The rest of the functions should keep on working in harmony without bugs for as long as all variables belong to the same inertial frame of reference. The only change a left-handed and right-handed system will have is that that graphical visualizations using a right handed system will show a mirror image of what will be observed in carla. This concludes this commit.

TODO:
1. Testing the API for multiple obstacles, both static and dynamic.
2. Integrating the API with a synchoronous carla simultion for single static obstacle.
3. Confirming euler angle handling order of carla for future implementations of better coordinate transforms.

fix: no fix made in this commit.
bugs: <carried forward from the last commit>