major feature/refactor: Towards ROS and Asynchronicity
This commit marks the beginning of adding a layer of ROS compatibility for the CARLA-ROS bridge to the cbf library. There are several other features to be added which will be added as feature requests in the near future. Therefore, please keep an eye out for issues tagged as feature requests.

bugs: #2 and #3

fixes: none

feature requests: none, all important feature requests will be added soon.

---- This concludes this major refactor commit ----