feature (WIP): graphical testbed for multiple dynamic obstacles
The idea is to create a systematic matplotlib animation featuring multiple circular obstacles of two types:
   - 1. Seekers: Obstacles which follow the ego vehicle in order to purposedly collide with the ego vehicle.
   - 2. Stationary: Self explanatory stationary obstacles.

The obstacles (of any of the types listed above) will be spawned randomly in a pre-specified annular region around the ego_vehicle. The seekers, according to the default seeking law, will follow the obstacle by a proportional law which decides their velocity based on the position difference between the ego vehicle's COM and the obstacle's center. These obstacles will then be used for CBF solvers with multiple constraints for testing the capabilities of the CBF Safety Filter in ideal environments.

Progress:
  - 1. The obstacles spawner is completed. More features will be added as and when required.
  - 2. The spawn and update methods for the obstacles are done.

WIP:
  - 1. The requisite dunder methods for the spawner are remaining. They can be kept at a lower priority since they will only be required for debugging related purposes in Jupyter Notebook.
  - 2. The Animation function is remaining. It will be the main backbone of the simulation as it takes the frame number as the argument which in turn drives the timeframe. Therefore, the main simulation codes from this point onwards will be kept in the animation function which will be called by the matplotlib animation object.

----This concludes this yet another feature adding commit----