feature: Added the 2D Collision Cone obstacle type
NOTE: The attributes of the collision cone look very different from the Ellipse2D type for various reasons. This is unforseen and undesirable since this leads to incoherent structures of the different CBF obstacle representations which is bad for a genralizing API in the long term. This will need to be fixed in the near future.