feature: State Augmented biycle model
As planned, a new classed name SADBM_CBF_2DS, where the SADBM stands for State Augmented Dynamic Bicycle Model was added to cbf.py, it inherits all the important methods from the primitive DBM_CBF_2DS. This class removes the small angle approximation, but has two important points of note:
1. It numerically differentiates the beta_ref calculated from steer_ref to get the reference d(beta)/dt, which is the new control i/p for the new model.
2. It numerically integrates the CBF's d(beta)/dt output to get the commanded beta_cbf, as well as for it to be used as the state variable value for calculating the f_c and g_c vectors for the CBF constraints.

So far, all the numerical integration and differentation methods are trivial, first order methods, which are bound to be unstable for ill-conditioned data streams or higher order derivatives, if any (none at the moment, and none expected in the future). So, in the long run, it is better to design a differentiator and intergator algorithm for real time signals, similar to the simulink blocks of the same name. This will mark the next steps towards the making of the API.