major fix: Fixed the PolyLane class
There were many logical errors in the update and the polynomial fitting function. Furthermore, there was a sign error in the dy derivate function and the evaluate function was supposed to return the CBF value instead of the polynomial value which was a major source of error.

the update_by_curve_fit method somehow had a copy which was a class method. That class method was commented for now and will be removed in future commits. It is being kept for testing purposes at the moment.